Move GitUtils into java-client-utils module
