+++ client/java-client-utils/src/main/java/org/sonarsource/sonarlint/core/client.utils/package-info.java
@@ -1,5 +1,5 @@
 /*
- * SonarLint Core - Commons
+ * SonarLint Core - Java Client Utils
  * Copyright (C) 2016-2023 SonarSource SA
  * mailto:info AT sonarsource DOT com
  *
@@ -17,26 +17,8 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.sonarlint.core.commons.log;
+@ParametersAreNonnullByDefault
+package org.sonarsource.sonarlint.core.client.utils;
 
-import java.io.PrintWriter;
-import java.io.StringWriter;
+import javax.annotation.ParametersAreNonnullByDefault;
 
-/**
- * Allow to redirect SonarLint logs to a custom output on client side
- */
-public interface ClientLogOutput {
-
-  void log(String formattedMessage, Level level);
-
-  enum Level {
-    ERROR, WARN, INFO, DEBUG, TRACE;
-  }
-
-  static String stackTraceToString(Throwable t) {
-    var stringWriter = new StringWriter();
-    var printWriter = new PrintWriter(stringWriter);
-    t.printStackTrace(printWriter);
-    return stringWriter.toString();
-  }
-}
