+++ client/java-client-utils/src/test/test-repos/child-from-non-analyzed.zip
