+++ client/rpc-java-client/pom.xml
@@ -5,6 +5,7 @@
     <groupId>org.sonarsource.sonarlint.core</groupId>
     <artifactId>sonarlint-core-parent</artifactId>
     <version>10.0-SNAPSHOT</version>
+    <relativePath>../../pom.xml</relativePath>
   </parent>
   <artifactId>sonarlint-rpc-java-client</artifactId>
   <name>SonarLint Core - RPC Java Client</name>
